Create Leap Motion Listener
This should make a listener object to go with the other leap motion file
Again, may or may not be properly formatted
I added swipe and screentap gestures and a vector that keeps track of where you tapped, maybe we could use this to select options from the main screen? I was gonna make a similar thing to the interaction box to try to make button-like things?